fix(apps/client): HASHI-80 하시픽 공통 리뷰 반영
