+++ apps/client/src/features/restaurantList/components/index.ts
@@ -1,0 +1,3 @@
+export { RestaurantCard } from './RestaurantCard'
+export { RestaurantFilterBar } from './RestaurantFilterBar'
+export { RestaurantImageList } from './RestaurantImageList'
