+++ apps/client/src/features/restaurantList/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useInfiniteRestaurantList } from './useInfiniteRestaurantList'
+export { useRestaurantListPage } from './useRestaurantListPage'
