+++ apps/client/src/features/restaurantList/index.ts
@@ -2,10 +2,11 @@ export {
   CATEGORY_OPTIONS,
   DEFAULT_CATEGORY_OPTION,
   HASHI_PICK_SORT_OPTIONS,
-  MOCK_RESTAURANTS,
   RESTAURANT_LIST_PAGE_SIZE,
 } from './constants'
+export { RestaurantListPage } from './RestaurantListPage'
+export { MOCK_RESTAURANTS } from './mocks'
 export { useInfiniteRestaurantList } from './hooks'
-export { RestaurantCard } from './RestaurantCard'
-export { RestaurantFilterBar } from './RestaurantFilterBar'
+export { useRestaurantListPage } from './hooks'
+export { RestaurantCard, RestaurantFilterBar } from './components'
 export type { FilterOption, Restaurant } from './types'
