+++ apps/client/src/features/restaurantList/mocks/index.ts
@@ -1,0 +1,1 @@
+export { MOCK_RESTAURANTS } from './restaurantList.mock'
