+++ apps/client/src/pages/hashiPick/HashiPick.spec.md
@@ -43,10 +43,9 @@
 - [x] `초기화`는 현재 열린 sheet의 draft 값만 기본값으로 되돌립니다.
 - [x] 필터 BottomSheet는 X 버튼으로만 닫힙니다.
 - [x] RestaurantCard는 반복 렌더링되고 카드 클릭 시 식당 상세 route로 이동합니다.
-- [x] 식당 리스트는 최초 10개를 렌더링하고 하단 sentinel 진입 시 10개씩 추가 렌더링합니다.
-- [x] 식당 이미지는 최대 5개까지 가로 스크롤 리스트로 표시합니다.
-- [x] 식당 리스트 공통 UI는 `features/restaurantList`에 두고, route/page 조립과 navigation 상태는 `HashiPickPage`가 소유합니다.
-- [x] 식당 리스트 UI의 color/typography는 대응되는 HDS token utility를 우선 사용합니다.
+- [x] 식당 이미지는 가로 스크롤 리스트로 표시하고 이미지가 없으면 공통 `DefaultImage` fallback을 사용합니다.
+- [x] 식당 리스트는 mock data 25개를 10개 단위로 무한스크롤 렌더링합니다.
+- [x] sticky header는 `z-fixed` 토큰을 사용합니다.
 
 ## Data Dependencies
 
@@ -76,17 +75,21 @@
 - server state: none
 - derived state:
   - category filter label: `전체`일 때 `음식 장르 선택`, 그 외 selected label
+- owner:
+  - list state and handlers are owned by `useRestaurantListPage`
 
 ## UI Structure
 
 ```text
 HashiPickPage
-  Header
-  RestaurantFilterBar
-  RestaurantCard list
-    RestaurantImageList
-  FilterBottomSheet(sort)
-  FilterBottomSheet(category)
+  RestaurantListPage
+    Header
+    RestaurantFilterBar
+    RestaurantCard list
+      RestaurantImageList
+        DefaultImage fallback
+    FilterBottomSheet(sort)
+    FilterBottomSheet(category)
 ```
 
 ## Component Mapping
@@ -98,15 +101,19 @@ HashiPickPage
   - `Button`
 - app shared component:
   - `FilterBottomSheet`
-- page-local component: none
 - feature component:
+  - `RestaurantListPage`
   - `RestaurantFilterBar`
   - `RestaurantCard`
   - `RestaurantImageList`
 - feature hook:
+  - `useRestaurantListPage`
   - `useInfiniteRestaurantList`
+- feature mock:
+  - `mocks/restaurantList.mock.ts`
+- page-local component: none
 - icon:
-  - `BackIcon`
+- `BackIcon`
   - `TapDownIcon`
   - `CheckIcon`
   - `StarFillIcon`
@@ -124,7 +131,7 @@ HashiPickPage
 
 ## Verification
 
-- [x] `corepack pnpm --filter @hashi/client test -- HashiPickPage.test.tsx FilterBottomSheet.test.tsx`
+- [x] `corepack pnpm --filter @hashi/client test -- HashiPickPage.test.tsx PopularRestaurantsPage.test.tsx FilterBottomSheet.test.tsx`
 - [x] `corepack pnpm --filter @hashi/client lint`
 - [x] `corepack pnpm --filter @hashi/client typecheck`
 - [x] `corepack pnpm --filter @hashi/client build`
