+++ apps/client/src/shared/components/defaultImage/index.ts
@@ -1,0 +1,2 @@
+export { DefaultImage } from './DefaultImage'
+export type { DefaultImageLogoSize, DefaultImageProps } from './DefaultImage'
